refactor(identity-core): Introspector 스펙을 코어로 이동
- 패키지 명을 introspect로 변경함
