--- identity/identity-client/src/main/java/com/seatliberator/seatliberator/identity/client/introspector/Introspector.java
@@ -1,5 +1,0 @@
-package com.seatliberator.seatliberator.identity.client.introspector;
-
-public interface Introspector {
-    com.seatliberator.seatliberator.identity.core.introspection.Introspection introspect(String token);
-}
