+++ identity/identity-core/src/main/java/com/seatliberator/seatliberator/identity/core/introspect/Introspection.java
@@ -1,4 +1,4 @@
-package com.seatliberator.seatliberator.identity.core.introspection;
+package com.seatliberator.seatliberator.identity.core.introspect;
 
 import com.seatliberator.seatliberator.identity.core.actor.Actor;
 import org.jspecify.annotations.Nullable;
