+++ identity/identity-core/src/main/java/com/seatliberator/seatliberator/identity/core/introspect/Introspector.java
@@ -1,0 +1,5 @@
+package com.seatliberator.seatliberator.identity.core.introspect;
+
+public interface Introspector {
+    Introspection introspect(String token);
+}
