+++ identity/identity-core/src/test/java/com/seatliberator/seatliberator/identity/core/introspect/SimpleIntrospectionTest.java
@@ -1,4 +1,4 @@
-package com.seatliberator.seatliberator.identity.core.introspection;
+package com.seatliberator.seatliberator.identity.core.introspect;
 
 import com.seatliberator.seatliberator.identity.core.actor.Actor;
 import com.seatliberator.seatliberator.identity.core.actor.SimpleActor;
